refactor: Change some Participant class methods to Player class.
- 플레이어에게 있으면 되는 기능들을 Participant 에서 플레이어로 이동
- checkBurst -> isBurst 변경
