+++ src/main/java/domain/user/Dealer.java
@@ -21,4 +21,5 @@ public class Dealer {
     public List<Card> getCards() {
         return cards;
     }
+
 }
